+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Optimized Money struct size. This was 18 bytes (padded 24 bytes), but is now 16 bytes (padded 16 bytes) (33% less).
   This means Money is the same size as Decimal struct!!!
-- Improved Parsing: memory allocation (75% less) and performance (2x) by removing Regex
+- Improved Parsing: memory allocation (75% less) and performance (2x)
 
 ### Removed
 -